FileService.CalculateFileChunkInfos()

+++ src/Syncronaut.Core/Models/FileChunkInfo.cs
@@ -8,11 +8,12 @@ namespace Syncronaut.Core.Models;
 
 public class FileChunkInfo
 {
-    public ulong Index { get; private set; }
-    public uint ChunkSize { get; private set; }
+    public int Index { get; private set; }
+    public int ChunkSize { get; private set; }
     public string Hash { get; private set; }
 
-    public FileChunkInfo(ulong index, uint chunkSize, string hash)
+    //public FileChunkInfo(ulong index, uint chunkSize, string hash)
+    public FileChunkInfo(int index, int chunkSize, string hash)
     {
         Index = index;
         ChunkSize = chunkSize;
